### 2021-12-09 #3986 @danielskatz
👋 @martinfleis - you've been suggested as the editor - are you able to do this?

A first step would be to do some minor proofreading of the paper and the references, including missing DOIs
