### 2021-12-13 #3986 @martinfleis
👋 Hi @ni1o1,

if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).

Can you please check the suggestions on references listed in https://github.com/openjournals/joss-reviews/issues/3986#issuecomment-989993592 and add missing DOIs (if available)?

Also, could you try to format the list starting on a line 35 of the manuscript as an actual markdown list? The syntax should look like this:

```md
- First item
- Second item
- Third item
- Fourth item
```
