+++ paper.bib
@@ -106,6 +106,7 @@
   title={Taxi Global Positioning System Data in Urban Road Network: A Methodology to Identify Key Road Clusters Based on Travel Speed--Traffic Volume Correlation},
   author={Li, Yi and Li, Weifeng and Yu, Qing and Yang, Han},
   journal={Transportation Research Record},
+  doi={10.1177/03611981211036684},
   pages={03611981211036684},
   year={2021},
   publisher={SAGE Publications Sage CA: Los Angeles, CA}
@@ -115,6 +116,7 @@
   title={Partitioning urban road network based on travel speed correlation},
   author={Yu, Qing and Li, Weifeng and Yang, Dongyuan and Zhang, Haoran},
   journal={International Journal of Transportation Science and Technology},
+  doi={10.1016/j.ijtst.2021.01.002},
   volume={10},
   number={2},
   pages={97--109},
