### 2021-12-14 0c5ec3e
updater quality
A english version function for preparation
