+++ src/transbigdata/quality.py
@@ -71,3 +71,46 @@ def data_summary(data,col = ['Vehicleid','Time'],show_sample_duration = False,ro
         print('上四分位 :',round(sd['duration'].quantile(0.75),roundnum),'秒')
         print('中位数 :',round(sd['duration'].quantile(0.5),roundnum),'秒')
         print('下四分位 :',round(sd['duration'].quantile(0.25),roundnum),'秒')
+
+
+
+def data_summary_en(data,col = ['Vehicleid','Time'],show_sample_duration = False,roundnum = 4):
+    '''
+    English version
+    输入数据，打印数据概况
+    
+    输入
+    -------
+    data : DataFrame
+        轨迹点数据
+    col : List
+        列名，按[个体ID，时间]的顺序
+    show_sample_duration : bool
+        是否输出个体采样间隔信息
+    roundnum : number
+        小数点取位数
+    '''
+    [Vehicleid,Time] = col
+    print('Amount of data')
+    print('-----------------')
+    print('Total number of data items:',len(data))
+    Vehicleid_count = data[Vehicleid].value_counts()
+    print('Total number of individuals:',len(Vehicleid_count))
+    print('Mean value of individual data items:',round(Vehicleid_count.mean(),roundnum))
+    print('Upper quartile of individual data items:',round(Vehicleid_count.quantile(0.75),roundnum))
+    print('Median volume of individual data items:',round(Vehicleid_count.quantile(0.5),roundnum))
+    print('Lower quartile of individual data items:',round(Vehicleid_count.quantile(0.25),roundnum))
+    print('')
+    print('Data time period')
+    print('-----------------')
+    print('Start time:',data[Time].min())
+    print('End time:',data[Time].max())
+    print('')
+    if show_sample_duration:
+        sd = sample_duration(data, col=[Vehicleid, Time])
+        print('Individual sampling interval')
+        print('-----------------')
+        print('Mean value:',round(sd['duration'].mean(),roundnum),'seconds')
+        print('Upper quartile:',round(sd['duration'].quantile(0.75),roundnum),'seconds')
+        print('Median:',round(sd['duration'].quantile(0.5),roundnum),'seconds')
+        print('Lower quartile:',round(sd['duration'].quantile(0.25),roundnum),'seconds')
