+++ requirements.txt
@@ -1,4 +1,5 @@
 geopandas
 matplotlib
+scipy
 plot_map >= 0.3.4
 CoordinatesConverter>=0.1.4
