### 2021-12-14 c47b557
Update example-taxi.rst
