+++ docs/source/example-taxi/example-taxi.rst
@@ -2,6 +2,7 @@
 ==============
 
 | 这个案例的Jupyter notebook: `点击这里 <https://github.com/ni1o1/transbigdata/blob/main/example/Example%201-Taxi%20GPS%20data%20processing.ipynb>`__.
+| 可以点击`这个链接<https://mybinder.org/v2/gh/ni1o1/transbigdata/9507de936806c34a4befd74aa9227b012569a6aa?urlpath=lab%2Ftree%2Fexample%2FExample%201-Taxi%20GPS%20data%20processing.ipynb>`__在线编辑器中尝试
 | 使用示例中的样例数据集在github仓库中，链接为：https://github.com/ni1o1/transbigdata/tree/main/example
 | 下面我们介绍如何使用TransBigData包，调用其中的函数实现对出租车GPS数据的快速处理。
 | 首先我们引入TransBigData包，并读取数据:
