### 2021-12-14 c3e10df
Update example-taxi.rst
