### 2021-12-14 a1ca7cb
Update example-taxi.rst
