### 2021-12-14 e0b2a12
Update example-taxi.rst
