+++ docs/source/metromodel/metromodel.rst
@@ -3,6 +3,8 @@
 
 这个案例的Jupyter notebook: `点击这里 <https://github.com/ni1o1/transbigdata/blob/main/example/Example%203-Modeling%20for%20subway%20network%20topology.ipynb>`__.
 
+| 可以点击 `这个链接 <https://mybinder.org/v2/gh/ni1o1/transbigdata/9507de936806c34a4befd74aa9227b012569a6aa?urlpath=lab%2Ftree%2Fexample%2FExample%203-Modeling%20for%20subway%20network%20topology.ipynb>`__ 在线编辑器中尝试
+
 下面的案例展示如何用TransBigData包抓取地铁线路，并构建地铁线网的拓扑网络模型
 
 爬取地铁线路
