### 2021-12-14 4979783
Update example-taxi.rst
