+++ requirements.txt
@@ -1,6 +1,7 @@
 rtree
 geopandas
 matplotlib
+seaborn
 scipy
 plot_map >= 0.3.4
 CoordinatesConverter>=0.1.4
