### 2021-12-16 #3986 @martinfleis
👋👋👋 Dear @davidbailey, @jGaboardi, and @anitagraser,

Would you be willing to assist in reviewing this submission for JOSS (Journal of Open Source Software)?

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "TransBigData: A Python package for transportation spatio-temporal big data processing, analysis and visualization". You can find more information at the top of this Github issue (#3986).

Thank you!
