+++ requirements.txt
@@ -6,3 +6,5 @@ scipy
 plot_map >= 0.3.4
 CoordinatesConverter>=0.1.4
 networkx
+keplergl
+igraph
