### 2021-12-18 186b731
Update plot_map.rst
