+++ README.md
@@ -246,10 +246,16 @@ plt.show()
     
 ![png](images/output_7_0.png)
     
-### Visualize Taxi GPS trajectories(with keplergl)
+### Visualize trajectories (with keplergl)
 
 ![gif](images/tbdexample1.gif)
+
+### Visualize data distribution (with keplergl)
+
 ![gif](images/tbdexample2.gif)
+
+### Visualize OD (with keplergl)
+
 ![gif](images/tbdexample3.gif)
 
 ## Related Links
