### 2021-12-21 #3986 @anitagraser
I haven't had chance to look at the code yet. Concerning the paper, I think there is one point in the [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html) that is currently not addressed:

> A description of how this software compares to other commonly-used packages in this research area.
