### 2021-12-21 #4021 @whedon
[REVIEW]: TransBigData: A Python package for transportation spatio-temporal big data processing, analysis and visualization
**Submitting author:** @ni1o1 (<a href="http://orcid.org/0000-0003-2513-2969">Qing Yu</a>)
**Repository:** https://github.com/ni1o1/transbigdata
**Branch with paper.md** (empty if default branch):
**Version:** v0.3.7
