### 2021-12-21 #4021 @martinfleis
We aim for reviews to be completed within about 2-4 weeks but considering upcoming holidays, feel free to start whenever it works for you. Please let me know if any of you require significantly more time. We can also use Whedon (our bot) to set automatic reminders if you know you'll be away for a known period of time.
