### 2021-12-21 #4021 @martinfleis
Moving here a comment by @anitagraser from the pre-review (https://github.com/openjournals/joss-reviews/issues/3986#issuecomment-998799093)

> I haven't had chance to look at the code yet. Concerning the paper, I think there is one point in the review criteria that is currently not addressed:

> > A description of how this software compares to other commonly-used packages in this research area.
