+++ src/transbigdata/traj.py
@@ -2,9 +2,134 @@ import geopandas as gpd
 import pandas as pd
 from shapely.geometry import Polygon,Point
 from .grids import GPS_to_grids,grids_centre
-from .preprocess import id_reindex
+from .preprocess import id_reindex,clean_same
 import math 
 import numpy as np
+
+def plot_activity(data,col = ['stime','etime','LONCOL','LATCOL']):
+    '''
+    输入个体的活动数据（单一个体），绘制活动图
+    
+    输入
+    ----------------
+    data : DataFrame
+        活动数据集
+    col : List
+        列名，分别为[活动开始时间，活动结束时间，活动所在栅格经度编号，活动所在栅格纬度编号]
+    '''
+    stime,etime,LONCOL,LATCOL = col
+    activity = data.copy()
+    activity['date'] = activity[stime].dt.date
+    dates = list(activity['date'].astype(str).drop_duplicates())
+    #扩充天
+    dates_all = []
+    minday = min(dates)
+    maxday = max(dates)
+    import datetime
+    thisdate = minday
+    while thisdate != maxday:
+        dates_all.append(thisdate)
+        thisdate = str((pd.to_datetime(thisdate+' 00:00:00')+datetime.timedelta(days =1)).date())
+    dates = dates_all
+    #计算活动
+    import matplotlib.pyplot as plt
+    import numpy as np
+    #计算活动持续时间
+    activity['duration'] = (activity[etime]-activity[stime]).dt.total_seconds()
+    activity = activity[-activity['duration'].isnull()]
+    #计算开始结束时间戳
+    import time
+    activity['ststmp'] = activity[stime].astype(str).apply(lambda x:time.mktime(time.strptime(x,'%Y-%m-%d %H:%M:%S'))).astype('int64')
+    activity['etstmp'] = activity[etime].astype(str).apply(lambda x:time.mktime(time.strptime(x,'%Y-%m-%d %H:%M:%S'))).astype('int64')
+    #提取活动信息
+    activityinfo = activity[[LONCOL,LATCOL]].drop_duplicates()
+    indexs = list(range(1,len(activityinfo)+1))
+    np.random.shuffle(indexs)
+    activityinfo['index'] = indexs
+    #定义活动点颜色
+    import matplotlib as mpl
+    norm = mpl.colors.Normalize(vmin=0,vmax=len(activityinfo))
+    from matplotlib.colors import ListedColormap
+    import seaborn as sns
+    cmap = ListedColormap(sns.hls_palette(n_colors=len(activityinfo), l=.5, s=0.8))
+    #绘制活动点分布
+    fig     = plt.figure(1,(10,5),dpi = 250)    
+    ax      = plt.subplot(111)
+    plt.sca(ax)
+    for day in range(len(dates)):
+        plt.bar(day,height = 24*3600,bottom= 0,width=0.4,color = (0,0,0,0.1))
+        stime = dates[day]+' 00:00:00'
+        etime = dates[day]+' 23:59:59'
+        bars = activity[(activity['stime']<etime) & (activity['etime']>stime)].copy()
+        bars['ststmp'] = bars['ststmp'] - time.mktime(time.strptime(stime,'%Y-%m-%d %H:%M:%S'))
+        bars['etstmp'] = bars['etstmp'] - time.mktime(time.strptime(stime,'%Y-%m-%d %H:%M:%S'))    
+        for row in range(len(bars)):
+            plt.bar(day,height = bars['etstmp'].iloc[row]-bars['ststmp'].iloc[row],
+                    bottom=bars['ststmp'].iloc[row],
+                    color = cmap(norm(activityinfo[(activityinfo[LONCOL] == bars[LONCOL].iloc[row])&
[…]
