+++ docs/source/conf.py
@@ -22,8 +22,8 @@ copyright = '2021, Qing Yu'
 author = 'Qing Yu'
 
 # The full version, including alpha/beta/rc tags
-release = '0.2.7'
-version = '0.2.7'
+release = '0.3.0'
+version = '0.3.0'
 html_logo = "_static/logo-wordmark-light.png"
 html_favicon = '_static/logo2.ico'
 # -- General configuration ---------------------------------------------------

[…]
