### 2022-01-06 ni1o1/transbigdata#4 @jGaboardi
[JOSS] State of the field?
Currently no other similar software packages are mentioned in the manuscript, which is a requirement. There should be an overview of comparable packages, even/especially if they are not written in Python. Further, if no other packages exist this should be mentioned.

xref: [this comment](https://github.com/openjournals/joss-reviews/issues/4021#issuecomment-998864320) from @anitagraser.
