### 2022-01-06 ni1o1/transbigdata#5 @jGaboardi
[JOSS] Community Guidelines?
There are currently no Community Guidelines (*Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support*). Thorough community guidelines can be added [here](https://github.com/ni1o1/transbigdata/community), and also some more description can be given in the [Related Links](https://github.com/ni1o1/transbigdata#related-links) section in `README.md`.

xref: openjournals/joss-reviews#4021
