### 2022-01-06 ni1o1/transbigdata#6 @jGaboardi
[JOSS] Installation & Dependencies
In order to make it most clear for users to install `geopandas`, the link in [`README`](https://github.com/ni1o1/transbigdata#installation) and the install docs should be: https://geopandas.org/en/stable/getting_started.html#installation

Also, the following packages are used in the docs/tutorials and should be added to [dependencies](https://transbigdata.readthedocs.io/en/latest/getting_started.html#id3) list:
  * `CoordinatesConverter>=0.1.4`
  * `leuvenmapmatching`
  * `igraph` (optional)
  * `osmnx` (optional)
  * `plot_map`
  * `rtree`
  * `seaborn` (optional)
  * `shapely`

xref: openjournals/joss-reviews#4021
