### 2022-01-06 ni1o1/transbigdata#7 @jGaboardi
[JOSS] Testing?
I may be missing something, but there does not appear to be any actual CI testing being performed on `transbigdata`. It seems that [line 36](https://github.com/ni1o1/transbigdata/blob/f1683f57c1f1cadfb31cb731c379b705faeea4bd/.github/workflows/python-app.yml#L36) of `.github/workflows/python-app.yml` simply runs `pytest` on [`test_transbigdata.py`](https://github.com/ni1o1/transbigdata/blob/main/test_transbigdata.py), which merely imports the package (and also happens in [`.travis.yml`](https://github.com/ni1o1/transbigdata/blob/main/.travis.yml)). If I am incorrect in this assessment, please show where the actual CI testing is being performed.

xref: openjournals/joss-reviews#4021
