### 2022-01-06 ni1o1/transbigdata#8 @jGaboardi
[JOSS] target audience?
Regarding the **Statement of Need**, the docs and manuscript clearly state *what* problems are being solved, but there is no mentioned of *who* the target audience is.

This should be added to:
  * the actual manuscript
  * `README.md`
  * the docs site

For reference:
* **Documentation**
  *Do the authors clearly state what problems the software is designed to solve and who the target audience is?*
* **Software paper**
  *Does the paper have a section titled 'Statement of Need' that clearly states what problems the software is designed to solve and who the target audience is?*

xref: openjournals/joss-reviews#4021
