### 2022-01-07 #4021 @jGaboardi
@ni1o1 @anitagraser @martinfleis

I have completed my first thorough pass of `transbigdata` and the manuscript, and have opened several issues and a PR, mostly involving documentation. Once the minor language issues, etc. are fixed in the paper, I would comfortable checking off those boxes. However, a sticking point that will ultimately hold back JOSS acceptance is the apparent lack of any testing being performed (ni1o1/transbigdata#7). This must be addressed.
