+++ test_transbigdata.py
@@ -1,4 +1,0 @@
-def test_transbigdata():
-	import transbigdata as tbd
-	print('well done!')
-	return 666
