+++ .travis.yml
@@ -1,7 +1,0 @@
-language: python
-python:
-    - 3.6
-install:
-    - pip install -U transbigdata
-script: 
-	- pytest -v --color=yes --doctest-modules src/transbigdata
