### 2022-01-08 #4021 @anitagraser
> We have fixed the testing issue by adding github testing workflow. We are working on other issues.

As far as I can see from https://github.com/ni1o1/transbigdata/runs/4740487080?check_suite_focus=true there is only one (docstring) test.
