### 2022-01-08 ni1o1/transbigdata#6 @anitagraser
It would also be helpful to recommend a Python version.

I see that https://github.com/ni1o1/transbigdata/runs/4740487080?check_suite_focus=true sets up 3.6 but 3.6 support is deprecated by GeoPandas (https://github.com/geopandas/geopandas/commit/8c1e3531b8ed23e086ca2a3eede81e344a591a5d)
