### 2022-01-08 #4021 @anitagraser
What is the JOSS policy regarding other languages than English?

For example, this library still has Chinese in error messages:
![image](https://user-images.githubusercontent.com/590385/148656465-66c7eb13-e5f3-4d9b-aad1-d1251df117d4.png)

I imagine this would violate the documentation requirement "Functionality documentation: Is the core functionality of the software documented to a satisfactory level".
