### 2022-01-08 ni1o1/transbigdata#11 @anitagraser
[JOSS] Language of status and error messages
I've been going through `Example 1-Taxi GPS data processing.ipynb` and found that some messages are in Chinese:

![image](https://user-images.githubusercontent.com/590385/148656805-df118112-0504-42e3-966b-3ae000b193d4.png)

![image](https://user-images.githubusercontent.com/590385/148656816-850a03d0-62d5-4859-bd5c-632371fc478e.png)

Considering a potentially global user base, I think the messages should be in English.

xref: [openjournals/joss-reviews#4021](https://github.com/openjournals/joss-reviews/issues/4021)
