### 2022-01-08 ni1o1/transbigdata#11 @anitagraser
On a similar note, it would be nice to have Englisch examples in the notebooks, e.g. `Example 2-Identifying arrival and departure information from Bus GPS data.ipynb`

![image](https://user-images.githubusercontent.com/590385/148657384-67c5b96b-cacf-4d0e-96e5-4df0bd192109.png)
