### 2022-01-08 #4021 @jGaboardi
@anitagraser I noticed this, as well. I was not aware that JOSS had such a policy. Also there is an English language version of the docs. However, having all error messages in English does seem the way to go and would greatly benefit `transbigdata`.
