### 2022-01-08 ni1o1/transbigdata#12 @anitagraser
[JOSS] ValueError in Example 4 notebook
In `Example 4-pNEUMA trajectory dataset processing,ipynb`, I get the following error:

![image](https://user-images.githubusercontent.com/590385/148657986-c36346f4-43b7-4bc4-8969-a0328aae8889.png)

xref: [openjournals/joss-reviews#4021](https://github.com/openjournals/joss-reviews/issues/4021)
