+++ src/transbigdata/grids.py
@@ -5,22 +5,22 @@ import math
 import numpy as np
 def rect_grids(bounds,accuracy = 500):
     '''
-    生成研究范围内的方形栅格
+    Generate the rectangular grids in the bounds
 
-    输入
+    Parameters
     -------
     bounds : List
-        生成范围的边界，[lon1,lat1,lon2,lat2] (WGS84坐标系) 其中，lon1,lat1是左下角坐标，lon2,lat2是右上角坐标 
+        Create the bounds, [lon1, lat1, lon2, lat2](WGS84), where lon1 , lat1 are the lower-left coordinates, lon2 , lat2 are the upper-right coordinates
     accuracy : number
-        栅格大小（米）
+        Grid size (meter)
                                                
 
-    输出
+    Returns
     -------
     grid : GeoDataFrame
-        栅格的GeoDataFrame，其中LONCOL与LATCOL为栅格的编号，HBLON与HBLAT为栅格的中心点坐标 
+        Grids’ GeoDataFrame, LONCOL and LATCOL are the index of grids, HBLON and HBLAT are the center of the grids
     params : List
-        栅格参数(lonStart,latStart,deltaLon,deltaLat)，分别为栅格左下角坐标与单个栅格的经纬度长宽
+        Gridding parameters (lonStart,latStart,deltaLon,deltaLat), lonStart and latStart are the lower-left coordinates, deltaLon, deltaLat are the length and width of a single grid
     '''
     #导入math包  
     #划定栅格划分范围
@@ -78,20 +78,20 @@ def rect_grids(bounds,accuracy = 500):
 
 def grid_params(bounds,accuracy = 500):
     '''
-    栅格参数获取
+    Generate gridding params
 
-    输入
+    Parameters
     -------
     bounds : List
-        生成范围的边界，[lon1,lat1,lon2,lat2] (WGS84坐标系) 其中，lon1,lat1是左下角坐标，lon2,lat2是右上角坐标 
+        Bounds of the study area， [lon1, lat1, lon2, lat2](WGS84), where lon1 , lat1 are the lower-left coordinates, lon2 , lat2 are the upper-right coordinates 
     accuracy : number
-        栅格大小（米）
+        Grid size (meter)
                                                
 
-    输出
+    Returns
     -------
     params : List
-        栅格参数(lonStart,latStart,deltaLon,deltaLat)，分别为栅格左下角坐标与单个栅格的经纬度长宽
+        Gridding parameters (lonStart,latStart,deltaLon,deltaLat), lonStart and latStart are the lower-left coordinates, deltaLon, deltaLat are the length and width of a single grid
 
     Examples
     -------
@@ -116,23 +116,23 @@ def grid_params(bounds,accuracy = 500):
 
 def GPS_to_grids(lon,lat,params):
     '''
-    GPS数据对应栅格编号。输入数据的经纬度列与栅格参数，输出对应的栅格编号
+    Match the GPS data to the grids. The input is the columns of longitude, latitude, and the grids parameter. The output is the grid ID.
 
-    输入
+    Parameters
     -------
     lon : Series
-        经度列
+        The column of longitude
     lat : Series
-        纬度列
+        The column of latitude
     params : List
-        栅格参数(lonStart,latStart,deltaLon,deltaLat)，分别为栅格左下角坐标与单个栅格的经纬度长宽
-                                               
-    输出
+        Gridding parameters (lonStart,latStart,deltaLon,deltaLat), lonStart and latStart are the lower-left coordinates, deltaLon, deltaLat are the length and width of a single grid
+                                            
+    Returns
     -------
     LONCOL : Series
-        经度栅格编号列
+        The index of the grid longitude. The two columns LONCOL and LATCOL together can specify a grid.
[…]
