+++ src/transbigdata/getbusdata.py
@@ -10,21 +10,23 @@ from urllib import request
 
 def getadmin(keyword,ak,subdistricts = False):
     '''
-    输入关键词与高德ak，抓取行政区划gis
-    输入
+    Input the keyword and the Amap ak. The output is the GIS file of the administrative boundary (Only in China)
+    
+    Parameters
     -------
     keywords : str
-        关键词，可以是名称，如"深圳市"，或行政区划编号，如440500
+        The keyword. It might be the city name such as Shengzheng, or the administrative code such as 440500
     ak : str
-        高德ak
+        Amap accesstoken
     subdistricts : bool
-        是否输出子行政区划的信息
-    输出
+        Whether to output the information of the administrative district boundary
+
+    Returns
     -------
     admin : GeoDataFrame
-        行政区划信息
+        Administrative district
     districts : DataFrame
-        子行政区划的信息，利用这个可以进一步抓下一级的行政区划
+        The information of subdistricts. This can be used to further get the boundary of lower level districts
     '''
     
     #查询的接口地址
@@ -110,28 +112,94 @@ def getadmin(keyword,ak,subdistricts = False):
 
 def getbusdata(city,keywords):
     '''
-    通过输入城市与关键词，获取公交线路的线型与站点
-    输入
+    Obtain the geographic information of the bus station and bus line from the map service (Only in China)
+
+    Parameters
     -------
     city : str
-        城市
+        city name
     keywords : List
-        关键词，线路名称
+        Keyword, the line name
 
-    输出
+    Returns
     -------
     data : GeoDataFrame
-        生成的公交线路
+        The generated bus line
     stop : GeoDataFrame
-        生成的公交站点
+        The generated bus station
     '''
-    print('获取城市id:',city,end = '')
+    def getlineuid(keyword,c):
+        url = 'http://map.baidu.com/?qt=s&wd='+urllib.parse.quote(keyword)+'&c='+c
+        response1 = urllib.request.urlopen(url)
+        searchinfo=json.loads(response1.read().decode('utf8'))
+        if searchinfo['content'][0]['catalogID'] ==904 or searchinfo['content'][0]['catalogID'] ==905:
+            try:
+                uidlist = list(pd.DataFrame(searchinfo['content'][8]['blinfo'])['uid'])
+            except:
+                uidlist = []
+            uidlist.append(searchinfo['content'][0]['uid'])
+            uidlist.append(searchinfo['content'][1]['uid'])
+            return list(set(uidlist))
+        else:
+            return []
+    def getcitycode(c):
+        url = 'http://map.baidu.com/?qt=s&wd='+urllib.parse.quote(c)
+        response1 = urllib.request.urlopen(url,timeout = 60)
+        searchinfo=json.loads(response1.read().decode('utf8'))
+        return str(searchinfo['content']['code'])
+    def getlinegeo(uid,c):
+        url = 'http://map.baidu.com/?qt=bsl&uid='+uid+'&c='+c
+        response = urllib.request.urlopen(url,timeout = 60)
+        searchinfo=json.loads(response.read().decode('utf8'))
+        linename = searchinfo['content'][0]['name']
+        stations = searchinfo['content'][0]['stations']
+        geo = searchinfo['content'][0]['geo'].split('|')[2][:-1].split(',')
+        stationgeo = []
+        stationnames = []
+        for station in stations:
+            stationname = station['name']
+            coo = station['geo'].split(';')[1].split('|')[0]
[…]

[…]
