### 2022-01-09 ni1o1/transbigdata#12 @ni1o1
I ran `Example 4-pNEUMA trajectory dataset processing.ipynb` again but I didn't encounter the above problem. This error happens when the input GeoDataFrame `tmpgdf_data` is empty, perhaps you lose the content of `tmpgdf_data` in previous steps.
