### 2022-01-09 ni1o1/transbigdata#12 @anitagraser
It looks like there is no track_id 2138 in the provided dataset. The maximum in gdf_data is 157:

![image](https://user-images.githubusercontent.com/590385/148679271-a872b563-3cb5-463d-9d6a-463868000c27.png)
