### 2022-01-09 ni1o1/transbigdata#12 @yuanjian24
Hi there, according to your picture, `gdf_data` only has 188 rows. However, the right data should contain 10674 rows of data. Please make sure that you got the right dataset by reading this file:
```
data = pd.read_csv('data/pNEUMA_tbd_sample.csv')
```
and run every cell before the mentioned cell.

![image](https://user-images.githubusercontent.com/50770426/148680027-016ac654-7d38-4e7c-9063-bc2238088875.png)

Or, you can also directly replace the `tmp_gdf_data` with `gdf_data`:

```
# get the nearest node of each point on the map
gdf_data = tbd.ckdnearest_point(gdf_data, nodes)

# extract the o/d node
o_index, d_index = gdf_data.iloc[0]['index'], gdf_data.iloc[-1]['index']
o_node_id, d_node_id = list(nodes[nodes['index']==o_index].index)[0], \
                       list(nodes[nodes['index']==d_index].index)[0]
print(o_node_id, d_node_id)
gdf_data.head()
```
![image](https://user-images.githubusercontent.com/50770426/148679966-ddf38bdc-e258-4512-8297-d0975500094a.png)
