### 2022-01-09 ni1o1/transbigdata#6 @anitagraser
I had to explicitly conda install python-igraph to make `import igraph` work.
