### 2022-01-09 ni1o1/transbigdata#12 @anitagraser
I just pulled the lastest version from the repo and the issue remains.

`gdf_data` is created from `data_sparsify_clean` which only has 128 rows:

```gdf_data = gpd.GeoDataFrame(data_sparsify_clean, ```

This has to be fixed in the notebook.
