### 2022-01-09 ni1o1/transbigdata#12 @yuanjian24
This is a little bit weird, I have double-checked the code. To solve the problem, I have made the `data_sparsify` dataframe to a CSV file: `data/data_sparsify.csv`. Please pull again and run all the cells again to see if it works.

![image](https://user-images.githubusercontent.com/50770426/148681464-e7cc7d53-bfe6-49a7-bc83-2997ab7e80e5.png)

Use `data_sparsify` to test `tbd.ckdnearest_point` if it still doesn't work
