### 2022-01-09 #4021 @anitagraser
@ni1o1 @martinfleis @jGaboardi

I have also finished my first pass through the paper, repo, and example notebooks. I've created new issues and added comments to @jGaboardi's issues where I noticed additional aspects. From my perspective, the lack of tests is the main issue with the repo. The lack of state of the art is the main issue with the paper.
