### 2022-01-09 ni1o1/transbigdata#12 @anitagraser
Rerunning the notebook over-writes your provided csv. So there should probably be a check if the file already exists.

With your file, the following cells run.

One strange effect is that I get the following plot:
![image](https://user-images.githubusercontent.com/590385/148682218-33b6dc39-b3de-4db7-8b0c-2b21a84a111d.png)

when the checked in version shows:
![image](https://user-images.githubusercontent.com/590385/148682239-f77caeb0-935a-47eb-a338-517424109a46.png)
