### 2022-01-09 #4021 @martinfleis
Thanks both, @anitagraser and @jGaboardi!

@ni1o1 I am also a bit concerned about the fact that docstrings, error messages and comments are in Chinese and would like to see it in English. However, I am trying to figure out whether it is an issue hindering publication with the wider editorial team. I'll get back to you on that.

Also, please let us know how you plan to respond to the comments raised by reviewers and what is the expected timeline for that. Thanks!
