### 2022-01-09 ni1o1/transbigdata#12 @yuanjian24
Hi, I have updated the e.g. 4 notebook:

The file `data/data_sparsify.csv` will not be overwritten now, and I added a file check and a check-points. You can start from here:
![image](https://user-images.githubusercontent.com/50770426/148684767-03c39e63-8006-4e80-b5aa-8e171fe42097.png)

Regarding the second problem, a `GeoDataFrame` format is required for visualization (with the column of 'geometry')
The code in the red box can print the list of vehicles in the sequence of points number, you can choose any number or follow my codes:
![image](https://user-images.githubusercontent.com/50770426/148684919-1a8c0d07-9b1f-4da5-9e38-8579d0680f52.png)
