### 2022-01-09 #4021 @ni1o1
> Thanks both, @anitagraser and @jGaboardi!
>
> @ni1o1 I am also a bit concerned about the fact that docstrings, error messages and comments are in Chinese and would like to see it in English. However, I am trying to figure out whether it is an issue hindering publication with the wider editorial team. I'll get back to you on that.
>
> Also, please let us know how you plan to respond to the comments raised by reviewers and what is the expected timeline for that. Thanks!

@martinfleis @anitagraser @jGaboardi Thank you all for reviewing our package!
Here is how we plan to improve our package according to the comments:
1. Translate all docstrings, error messages and comments into English.
2. Add necessary tests into the repo.
3. Add state of the art and state of the target audience into the paper.
4. Other improvment of the docs and examples.
We plan to finish them within three weeks.
