### 2022-01-09 ni1o1/transbigdata#6 @jGaboardi
@anitagraser `conda install igraph` did not work for me (due to what you mentioned above), but `pip install igraph` did work correctly.
