### 2022-01-09 ni1o1/transbigdata#12 @jGaboardi
@anitagraser hmmm. I did not have any issues with the original `Example 4-pNEUMA trajectory dataset processing.ipynb` and could reproduce the notebook.
