### 2022-01-09 ni1o1/transbigdata#12 @anitagraser
Might be a local issue then. I'm testing on Win10 with Conda.

In my projects, I find it helpful to set up the notebooks so they can be tested on MyBinder. It gives some extra assurance that the notebooks also work on other machines than my own.
