+++ paper.bib
@@ -123,3 +123,83 @@
   year={2021},
   publisher={Elsevier}
 }
+
+@article{2020moveVis,
+author = {Schwalb-Willmann, Jakob and Remelgado, Ruben and Safi, Kamran and Wegmann, Martin},
+title = {moveVis: Animating movement trajectories in synchronicity with static or temporally dynamic environmental data in r},
+journal = {Methods in Ecology and Evolution},
+volume = {11},
+number = {5},
+pages = {664-669},
+keywords = {animal tracking, animation, data visualization, movement data, movement ecology, moveVis, r, spatio-temporal data},
+doi = {https://doi.org/10.1111/2041-210X.13374},
+url = {https://besjournals.onlinelibrary.wiley.com/doi/abs/10.1111/2041-210X.13374},
+eprint = {https://besjournals.onlinelibrary.wiley.com/doi/pdf/10.1111/2041-210X.13374},
+abstract = {Abstract Visualizing movement data is challenging: While traditional spatial data can be sufficiently displayed as two-dimensional plots or maps, movement trajectories require the representation of time in a third dimension. To address this, we present moveVis, an r package, which provides tools to animate movement trajectories, overlaying simultaneous uni- or multi-temporal raster imagery or vector data. moveVis automates the processing of movement and environmental data to turn such into an animation. This includes (a) the regularization of movement trajectories enforcing uniform time instances and intervals across all trajectories, (b) the frame-wise mapping of movement trajectories onto temporally static or dynamic environmental layers, (c) the addition of customizations, for example, map elements or colour scales and (d) the rendering of frames into an animation encoded as GIF or video file. moveVis is designed to display interactions and concurrencies of animal movement and environmental data. We present examples and use cases, ranging from data exploration to visualizing scientific findings. Static spatial plots of movement data disregard the temporal dimension that distinguishes movement from other spatial data. In contrast, animations allow to display relocation in both time and space. We deem animations a powerful way to visually explore movement data, frame analytical findings and display potential interactions with spatially continuous and temporally dynamic environmental covariates.},
+year = {2020}
+}
+
+@article{Siebinga2021,
+  doi = {10.21105/joss.03607},
+  url = {https://doi.org/10.21105/joss.03607},
+  year = {2021},
+  publisher = {The Open Journal},
+  volume = {6},
+  number = {65},
+  pages = {3607},
+  author = {Olger Siebinga},
+  title = {TraViA: a Traffic data Visualization and Annotation tool in Python},
+  journal = {Journal of Open Source Software}
+}
+@article{
+0x003aba2b,
+editor = {Adrijana Car},
+editor = {Thomas Jekel},
+editor = {Josef Strobl},
+editor = {Gerald Griesebner (Eds.)},
+subject = {geography},
[…]
