+++ setup.py
@@ -34,4 +34,6 @@ setuptools.setup(
     package_dir={"": "src"},
     packages=setuptools.find_packages(where="src"),
     python_requires=">=3.6",
+	test_suite='nose.collector',
+    tests_require=['nose'],
 )
