+++ .github/workflows/tests.yml
@@ -20,20 +20,24 @@ jobs:
       uses: actions/setup-python@v2
       with:
         python-version: "3.6"
+
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
         pip install flake8 pytest
         if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
+    
     - name: Lint with flake8
       run: |
         # stop the build if there are Python syntax errors or undefined names
         flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
         # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
+    
     - name: Test docstrings
       run: |
         pytest -v --color=yes --doctest-modules ./src/transbigdata
+    
     - name: Test 
       run: |
         pip install pytest-cov
