+++ src/transbigdata/tests/test_grids.py
@@ -4,58 +4,71 @@ import pandas as pd
 import geopandas as gpd
 from shapely.geometry import Polygon
 
+
 class TestGrids:
     def setup_method(self):
-        self.bounds = [113.6,22.4,113.605,22.405]
+        self.bounds = [113.6, 22.4, 113.605, 22.405]
         self.params = (113.6, 22.4, 0.004863669213934598, 0.004496605206422906)
 
     def test_rect_grids(self):
-        result = tbd.rect_grids(self.bounds,accuracy = 500)
+        result = tbd.rect_grids(self.bounds, accuracy=500)
         res1 = result[0]['geometry'].to_wkt().iloc[0]
         res2 = result[1]
-        assert res1=='POLYGON ((113.597568165393 22.39775169739679, 113.602431834607 22.39775169739679, 113.602431834607 22.40224830260321, 113.597568165393 22.40224830260321, 113.597568165393 22.39775169739679))'
+        assert res1 == 'POLYGON ((113.597568165393 22.39775169739679, 113.602431834607 22.39775169739679, 113.602431834607 22.40224830260321, 113.597568165393 22.40224830260321, 113.597568165393 22.39775169739679))'
         truth = (113.6, 22.4, 0.004863669213934598, 0.004496605206422906)
-        assert np.allclose(res2,truth)
+        assert np.allclose(res2, truth)
 
     def test_grid_params(self):
-        result = tbd.grid_params(self.bounds,accuracy = 500)
+        result = tbd.grid_params(self.bounds, accuracy=500)
         truth = (113.6, 22.4, 0.004863669213934598, 0.004496605206422906)
-        assert np.allclose(result,truth)
+        assert np.allclose(result, truth)
 
     def test_GPS_to_grids(self):
-        result = tbd.GPS_to_grids(113.7,22.7,self.params)
+        result = tbd.GPS_to_grids(113.7, 22.7, self.params)
         truth = (21, 67)
-        assert np.allclose(result,truth)
+        assert np.allclose(result, truth)
 
     def test_grids_centre(self):
-        result = tbd.grids_centre(21, 67,self.params)
+        result = tbd.grids_centre(21, 67, self.params)
         truth = (113.70213705349262, 22.70127254883033)
-        assert np.allclose(result,truth)
+        assert np.allclose(result, truth)
 
     def test_gridid_to_polygon(self):
-        result = tbd.gridid_to_polygon(pd.Series(21), pd.Series(67),self.params)[0].wkt
+        result = tbd.gridid_to_polygon(
+            pd.Series(21), pd.Series(67), self.params)[0].wkt
         truth = 'POLYGON ((113.6997052188857 22.69902424622712, 113.7045688880996 22.69902424622712, 113.7045688880996 22.70352085143355, 113.6997052188857 22.70352085143355, 113.6997052188857 22.69902424622712))'
-        assert result==truth
-    
+        assert result == truth
+
     def test_hexagon_grids(self):
-        result = tbd.hexagon_grids(self.bounds,accuracy = 500)['geometry'].iloc[0].wkt
-        truth = 'POLYGON ((113.5955084235794 22.4, 113.5977542117897 22.40359627046783, 113.6022457882103 22.40359627046783, 113.6044915764206 22.4, 113.6022457882103 22.39640363649211, 113.5977542117897 22.39640363649211, 113.5955084235794 22.4))'
-        assert result==truth
-    
+        result = tbd.hexagon_grids(self.bounds, accuracy=500)[
+            'geometry'].iloc[0]
+        result = np.array(result.exterior.coords)
+        truth = [[113.59550842,  22.4],
+                 [113.59775421,  22.40359627],
+                 [113.60224579,  22.40359627],
[…]
