+++ src/transbigdata/tests/test_gisprocess.py
@@ -1,0 +1,107 @@
+import transbigdata as tbd
+import numpy as np
+import pandas as pd
+import geopandas as gpd
+from shapely.geometry import LineString, Polygon
+
+
+class TestGisprocess:
+    def setup_method(self):
+        self.Centerline = gpd.read_file('{"type": "FeatureCollection", "features": [{"id": "0", "type": "Feature", "properties": {"Id": 0}, "geometry": {"type": "LineString", "coordinates": [[87.68570740160918, 44.0787457866043], [87.68645046729048, 44.076232872455286]]}}, {"id": "1", "type": "Feature", "properties": {"Id": 0}, "geometry": {"type": "LineString", "coordinates": [[87.68264271353992, 44.075612961425655], [87.68375615646332, 44.07414940028094]]}}, {"id": "2", "type": "Feature", "properties": {"Id": 0}, "geometry": {"type": "LineString", "coordinates": [[87.68645046729048, 44.076232872455286], [87.68708471544996, 44.07365060928234]]}}, {"id": "3", "type": "Feature", "properties": {"Id": 0}, "geometry": {"type": "LineString", "coordinates": [[87.68645046729048, 44.076232872455286], [87.6875136241044, 44.07504676107092], [87.68851961299117, 44.0746480974192]]}}, {"id": "4", "type": "Feature", "properties": {"Id": 0}, "geometry": {"type": "LineString", "coordinates": [[87.68645046729048, 44.076232872455286], [87.68621557677827, 44.07507660904865], [87.6856695946903, 44.0744984489369]]}}, {"id": "5", "type": "Feature", "properties": {"Id": 0}, "geometry": {"type": "LineString", "coordinates": [[87.68375615646332, 44.07414940028094], [87.68443421646374, 44.07310623751142], [87.68516775880424, 44.072626945592525], [87.68612967919796, 44.07240268063775], [87.68760135753165, 44.07243204581898]]}}, {"id": "6", "type": "Feature", "properties": {"Id": 0}, "geometry": {"type": "LineString", "coordinates": [[87.68375615646332, 44.07414940028094], [87.6856695946903, 44.0744984489369]]}}, {"id": "7", "type": "Feature", "properties": {"Id": 0}, "geometry": {"type": "LineString", "coordinates": [[87.68733133758964, 44.07337902470143], [87.68760135753165, 44.07243204581898]]}}, {"id": "8", "type": "Feature", "properties": {"Id": 0}, "geometry": {"type": "LineString", "coordinates": [[87.68708471544996, 44.07365060928234], [87.68733133758964, 44.07337902470143]]}}, {"id": "9", "type": "Feature", "properties": {"Id": 0}, "geometry": {"type": "LineString", "coordinates": [[87.68733133758964, 44.07337902470143], [87.68778577868343, 44.07424246258923], [87.68851961299117, 44.0746480974192]]}}]}')
+        self.dfA = gpd.GeoDataFrame([[1, 2], [2, 4], [2, 6],
+                                     [2, 10], [24, 6], [21, 6],
+                                     [22, 6]], columns=['lon1', 'lat1'])
+        self.dfB = gpd.GeoDataFrame(
+            [[1, 3], [2, 5], [2, 2]], columns=['lon', 'lat'])
+
+    def test_splitline_with_length(self):
+        splitedline = tbd.splitline_with_length(self.Centerline, maxlength=100)
+        result = splitedline['length'].max()
+        assert np.allclose(result, 0.0045634806283561345)
+
+    def test_ckdnearest(self):
+
[…]
