### 2022-01-10 58527f2
Create test_visualization.py
