### 2022-01-10 #4021 @martinfleis
> @ni1o1 I am also a bit concerned about the fact that docstrings, error messages and comments are in Chinese and would like to see it in English. However, I am trying to figure out whether it is an issue hindering publication with the wider editorial team. I'll get back to you on that.

Just for the sake of completeness, the agreed editorial position on this is that **the content (docstrings, comments, error messages) is subject to review and thus needs to be in the same language as the paper**.
