+++ src/transbigdata/tests/test_busdata.py
@@ -1,0 +1,121 @@
+
+import transbigdata as tbd
+import numpy as np
+import pandas as pd
+import geopandas as gpd
+from shapely.geometry import Polygon
+
+
+class TestBikedata:
+    def setup_method(self):
+        self.data = pd.DataFrame([['沪D-R7103', '2019-01-16 23:59:59', 121.3308582912764,
+                                  31.175129076427655],
+                                 ['沪D-R7103', '2019-01-17 12:40:20', 121.33049311941906,
+                                  31.175065802450753],
+                                 ['沪D-R7103', '2019-01-17 13:05:32', 121.33049311941906,
+                                  31.175065802450753],
+                                 ['沪D-R7103', '2019-01-17 13:05:52', 121.33036738445367,
+                                  31.175279932303557],
+                                 ['沪D-R7103', '2019-01-17 13:06:07', 121.32967297800323,
+                                  31.17497940724048],
+                                 ['沪D-R7103', '2019-01-17 13:06:22', 121.33008305045536,
+                                  31.174995618170627],
+                                 ['沪D-R7103', '2019-01-17 13:06:42', 121.33029256851628,
+                                  31.17510216522793],
+                                 ['沪D-R7103', '2019-01-17 16:24:05', 121.33029256851628,
+                                  31.17510216522793],
+                                 ['沪D-R7103', '2019-01-17 16:24:30', 121.33093714049436,
+                                  31.174819084849073],
+                                 ['沪D-R7103', '2019-01-17 16:25:00', 121.33119656323854,
+                                  31.174744630573823],
+                                 ['沪D-R7103', '2019-01-17 16:25:35', 121.3313133039642,
+                                  31.17470742816974],
+                                 ['沪D-R7103', '2019-01-17 16:34:29', 121.3313133039642,
+                                  31.17470742816974],
+                                 ['沪D-R7103', '2019-01-17 16:34:54', 121.33121258219445,
+                                  31.17410692572959],
+                                 ['沪D-R7103', '2019-01-17 16:35:20', 121.32927503207796,
+                                  31.173189080128747],
+                                 ['沪D-R7103', '2019-01-17 16:35:45', 121.32774259804339,
+                                  31.17211555668083],
+                                 ['沪D-R7103', '2019-01-17 16:36:05', 121.32828742144076,
+                                  31.171421870458047],
+                                 ['沪D-R7103', '2019-01-17 16:36:35', 121.32889308056538,
+                                  31.171053902452424],
+                                 ['沪D-R7103', '2019-01-17 16:37:00', 121.33147515535293,
+                                  31.172212396845183],
+                                 ['沪D-R7103', '2019-01-17 16:37:25', 121.33256465354401,
+                                  31.17289898466245],
+                                 ['沪D-R7103', '2019-01-17 16:38:35', 121.33256465354401,
+                                  31.17289898466245],
+                                 ['沪D-R7103', '2019-01-17 16:39:01', 121.33345761253085,
+                                  31.17344502270766],
+                                 ['沪D-R7103', '2019-01-17 16:39:21', 121.33442839412734,
[…]
