+++ src/transbigdata/busgps.py
@@ -1,8 +1,5 @@
 import geopandas as gpd  
 import pandas as pd
-from shapely.geometry import Polygon,Point
-from .grids import GPS_to_grids,grids_centre
-import math 
 import numpy as np
 from .preprocess import *
 
@@ -40,7 +37,7 @@ def busgps_arriveinfo(data,line,stop,col = ['VehicleId','GPSDateTime','lon','lat
         Bus arrival and departure information
     '''
     VehicleId,GPSDateTime,lon,lat,stopcol = col
-    #数据清洗
+    #Clean data
     print('Cleaning data',end = '')
     line.set_crs(crs='epsg:4326',allow_override=True,inplace=True)
     line = line.to_crs(epsg = project_epsg)
@@ -55,7 +52,7 @@ def busgps_arriveinfo(data,line,stop,col = ['VehicleId','GPSDateTime','lon','lat
     data = id_reindex(data,VehicleId,timegap = timegap,timecol = GPSDateTime,suffix='')
 
     print('Position matching',end = '')
-    #利用project方法，将数据点投影至公交线路上
+    #project data points onto bus LineString
     lineshp = line['geometry'].iloc[0]
     print('.',end = '')
     data['geometry'] = gpd.points_from_xy(data[lon],data[lat])
@@ -68,7 +65,7 @@ def busgps_arriveinfo(data,line,stop,col = ['VehicleId','GPSDateTime','lon','lat
         data['project'] = data['geometry'].apply(lambda r:lineshp.project(r))
     elif method == 'dislimit':
         tmps = []
-        #改进的匹配方法
+        #Distance limit method 
         for vid in data[VehicleId].drop_duplicates():
             print('.',end = '')
             tmp = data[data[VehicleId]==vid].copy()
@@ -96,41 +93,34 @@ def busgps_arriveinfo(data,line,stop,col = ['VehicleId','GPSDateTime','lon','lat
             tmps.append(tmp)
         data = pd.concat(tmps)
     print('.',end = '')
-    #公交站点也进行project
+    #Project bus stop to bus line
     stop = stop.to_crs(epsg = project_epsg)
     stop['project'] = stop['geometry'].apply(lambda r:lineshp.project(r))
     print('.',end = '')
-    #标准化时间
     starttime = data[GPSDateTime].min()
     data['time_st'] = (data[GPSDateTime]-starttime).dt.total_seconds()
     BUS_project = data
     print('.')
-    from shapely.geometry import LineString,Polygon
+    from shapely.geometry import LineString
     import shapely
-
-    #定义一个空的list存储识别结果
     ls = []
-
     print('Matching arrival and leaving info...',end = '')
-    #对每一辆车遍历
     for car in BUS_project[VehicleId].drop_duplicates():
         print('.',end = '')
-        #提取车辆轨迹
+        #Extract bus trajectory
         tmp = BUS_project[BUS_project[VehicleId] == car]
-        #如果车辆数据点少于1个，则无法构成轨迹
         if len(tmp)>1:
-            #对每一个站点识别
             for stopname in stop[stopcol].drop_duplicates():
-                #提取站点位置
+                #Get the stop position
                 position = stop[stop[stopcol] == stopname]['project'].iloc[0]
-                #通过缓冲区与线段交集识别到离站轨迹
+                #Identify arrival and departure by intersection of stop buffer and line segment
                 buffer_polygon = LineString([[0,position],
                                              [data['time_st'].max(),position]]).buffer(stopbuffer)
                 bus_linestring = LineString(tmp[['time_st','project']].values)
[…]
