### 2022-01-25 cd13fbc
Create CODE_OF_CONDUCT.md
