+++ README.md
@@ -263,5 +263,6 @@ plt.show()
 
 ## Introducing Video (In Chinese)
 
-[![Video1](https://i1.hdslb.com/bfs/archive/bc1ffd9264b22b97854d9b123643e2e00e3e5a5d.jpg@640w_400h_100Q_1c.webp)](https://www.bilibili.com/video/BV1nq4y1u7i1 "我写了一个Python包，时空大数据分析从此零门槛（TransBigData）")
+* [Bilibili](https://www.bilibili.com/video/BV1nq4y1u7i1)
+* [Youtube](https://www.youtube.com/watch?v=V_KHFv75W_w)
 
