### 2022-01-27 ad6ccd8
updata state of the art
fix a bug in Contributing to TransBigData @ni1o1
