+++ paper.md
@@ -27,13 +27,17 @@ bibliography: paper.bib
 ---
 # Summary
 
-In recent years, data generated in the field of transportation has begun to explode. Individual continuous tracking data, such as mobile phone data, IC smart card data, taxi GPS data, bus GPS data and bicycle sharing order data, also known as "spatio-temporal big data" or "Track &Trace data" [@harrison:2020], has great potential for applications in data-driven transportation research. These spatio-temporal big data typically require three aspects of information [@zhang:2021]: Who? When? Where? They are characterized by high data quality, large collection scope, and fine-grained spatio-temporal information, which can fully capture the daily activities of individuals and their travel behavior in the city in both temporal and spatial dimensions. The emergence of these data provides new ways and opportunities for potential transportation demand analysis and travel mechanism understanding in supporting urban transportation planning and management [@chen:2021; @zhang:2020]. However, processing with these multi-source spatial temporal big data usually requires a series of similar processing procedure (e.g., data quality accessment, data preprocessing, data cleaning, data gridding, data aggregation, and data visualization). There is an urgent need for a one-size-fits-all tool that can adapt to the various processing demand of different transportation data in this field.
+In recent years, data generated in the field of transportation has begun to explode. Individual continuous tracking data, such as mobile phone data, IC smart card data, taxi GPS data, bus GPS data and bicycle sharing order data, also known as "spatio-temporal big data" or "Track &Trace data" [@harrison:2020], has great potential for applications in data-driven transportation research. These spatio-temporal big data typically require three aspects of information [@zhang:2021]: Who? When? Where? They are characterized by high data quality, large collection scope, and fine-grained spatio-temporal information, which can fully capture the daily activities of individuals and their travel behavior in the city in both temporal and spatial dimensions. The emergence of these data provides new ways and opportunities for potential transportation demand analysis and travel mechanism understanding in supporting urban transportation planning and management [@chen:2021; @zhang:2020]. However, processing with these multi-source spatial temporal big data usually requires a series of similar processing procedure (e.g., data quality accessment, data preprocessing, data cleaning, data gridding, data aggregation, and data visualization). There is an urgent need for a one-size-fits-all tool that can adapt to the various processing demand of different transportation data in this field. 
 
 # **State of the art**
 
-Several similar tools exist which involves process and visualization of spatio-temporal big data. `moveVis` [@2020moveVis] providing tools to visualize movement data and temporal changes of environmental data by creating video animations. `TraViA` [@Siebinga2021] provides tools to visualize and annotate movement data in an interactive approach. `MovingPandas` [@0x003aba2b] is a comprehensive tool that could both provide trajectory data structures and functions for data exploration and visualization. `PySAL` [@rey_pysal_2010] is a family of packages allows for advanced geospatial data science, which supports the development of high-level applications. The traffic library [@Olive2019] is a toolbox for processing and analyzing air traffic data.
+Typical process for spatiotemporal data analysis include preprocessing, spatiotemporal analysis, and data visualization. Currently, there exist lot of packages in each individual or multiply domains: 
 
-Each of the softwares described above provide processing and analysis support from sepcific aspects of spatio-temporal big datatrack big data process. However, there are still a lot of works to do during the process of transforming raw spatio-temporal data into valuable insights. These tasks can hardly be fully covered by these softwares. Thus, a tool that integrates multiple data processing methods, GIS analysis tools, and various visualization features through one programming language could effectively facilitate the research progress.
[…]
