+++ paper.bib
@@ -1,3 +1,22 @@
+@article{haidri2021ptrail,
+      title={PTRAIL -- A python package for parallel trajectory data preprocessing},
+      author={Salman Haidri and Yaksh J. Haranwala and Vania Bogorny and Chiara Renso and Vinicius Prado da Fonseca and Amilcar Soares},
+      year={2021},
+      doi = {2108.13202},
+      eprint={2108.13202},
+      url={https://arxiv.org/abs/2108.13202},
+      archivePrefix={arXiv},
+      primaryClass={cs.DC}
+}
+
+@incollection{axhausen2007definition,
+  title = {Definition of movement and activity for transport modelling},
+  doi = {10.1108/9780857245670-016},
+  author = {Axhausen, Kay Werner},
+  booktitle = {Handbook of transport modelling},
+  year={2007},
+  publisher={Emerald Group Publishing Limited}
+}
 
 @article{yu:2020,
     title = {Mobile phone data in urban bicycle-sharing: {Market}-oriented sub-area division and spatial analysis on emission reduction potentials},
@@ -202,4 +221,3 @@ address = {Wien},
 	keywords = {Exploratory Spatial Data Analysis, Spatial Analytical Method, Spatial Data Analysis, Spatial Weight, Spatial Weight Matrice},
 	pages = {175--193},
 }
-
