### 2022-01-27 a689237
state of the art v2
add descriptions of geopandas and osmnx
