+++ paper.bib
@@ -1,3 +1,14 @@
+@article{boeing2017osmnx,
+  title={OSMnx: New methods for acquiring, constructing, analyzing, and visualizing complex street networks},
+  author={Boeing, Geoff},
+  doi = {10.1016/j.compenvurbsys.2017.05.004},
+  journal={Computers, Environment and Urban Systems},
+  volume={65},
+  pages={126--139},
+  year={2017},
+  publisher={Elsevier}
+}
+
 @article{haidri2021ptrail,
       title={PTRAIL -- A python package for parallel trajectory data preprocessing},
       author={Salman Haidri and Yaksh J. Haranwala and Vania Bogorny and Chiara Renso and Vinicius Prado da Fonseca and Amilcar Soares},
