+++ paper.bib
@@ -232,3 +232,17 @@ address = {Wien},
 	keywords = {Exploratory Spatial Data Analysis, Spatial Analytical Method, Spatial Data Analysis, Spatial Weight, Spatial Weight Matrice},
 	pages = {175--193},
 }
+@scikitmobility{pappalardo_luca_2019_5528110,
+  author       = {Pappalardo, Luca and
+                  Simini, Filippo and
+                  Barlacchi, Gianni and
+                  Pellungrini, Roberto},
+  title        = {scikit-mobility},
+  month        = jul,
+  year         = 2019,
+  note         = {If you use this code, please cite this software.},
+  publisher    = {Zenodo},
+  version      = {1.1.0},
+  doi          = {10.5281/zenodo.5528110},
+  url          = {https://doi.org/10.5281/zenodo.5528110}
+}
