+++ CITATION.cff
@@ -4,7 +4,7 @@ authors:
   - name: Qing Yu
     orcid: 0000-0003-2513-2969
     affiliation: 1
-  - name: Jian Yuan^[corresponding author]
+  - name: Jian Yuan
     orcid: 0000-0002-7202-0946
     affiliation: 1
 affiliations:
