+++ paper.bib
@@ -162,7 +162,7 @@ volume = {11},
 number = {5},
 pages = {664-669},
 keywords = {animal tracking, animation, data visualization, movement data, movement ecology, moveVis, r, spatio-temporal data},
-doi = {https://doi.org/10.1111/2041-210X.13374},
+doi = {doi.org/10.1111/2041-210X.13374},
 url = {https://besjournals.onlinelibrary.wiley.com/doi/abs/10.1111/2041-210X.13374},
 eprint = {https://besjournals.onlinelibrary.wiley.com/doi/pdf/10.1111/2041-210X.13374},
 abstract = {Abstract Visualizing movement data is challenging: While traditional spatial data can be sufficiently displayed as two-dimensional plots or maps, movement trajectories require the representation of time in a third dimension. To address this, we present moveVis, an r package, which provides tools to animate movement trajectories, overlaying simultaneous uni- or multi-temporal raster imagery or vector data. moveVis automates the processing of movement and environmental data to turn such into an animation. This includes (a) the regularization of movement trajectories enforcing uniform time instances and intervals across all trajectories, (b) the frame-wise mapping of movement trajectories onto temporally static or dynamic environmental layers, (c) the addition of customizations, for example, map elements or colour scales and (d) the rendering of frames into an animation encoded as GIF or video file. moveVis is designed to display interactions and concurrencies of animal movement and environmental data. We present examples and use cases, ranging from data exploration to visualizing scientific findings. Static spatial plots of movement data disregard the temporal dimension that distinguishes movement from other spatial data. In contrast, animations allow to display relocation in both time and space. We deem animations a powerful way to visually explore movement data, frame analytical findings and display potential interactions with spatially continuous and temporally dynamic environmental covariates.},
@@ -172,6 +172,7 @@ year = {2020}
 @article{Siebinga2021,
   doi = {10.21105/joss.03607},
   url = {https://doi.org/10.21105/joss.03607},
+  doi={10.21105/joss.03607},
   year = {2021},
   publisher = {The Open Journal},
   volume = {6},
@@ -181,24 +182,15 @@ year = {2020}
   title = {TraViA: a Traffic data Visualization and Annotation tool in Python},
   journal = {Journal of Open Source Software}
 }
-@article{
-0x003aba2b,
-editor = {Adrijana Car},
-editor = {Thomas Jekel},
-editor = {Josef Strobl},
-editor = {Gerald Griesebner (Eds.)},
-subject = {geography},
-abstract = {Movement data analysis is a high-interest topic in many scientific domains. Even though Python is the scripting language of choice in the GIS world, currently there is no Python library that would enable researchers and practitioners to interact with and analyse movement data efficiently. To close this gap, we present MovingPandas, a new Python library for dealing with movement data. Its development is based on an analysis of state-of-the-art conceptual frameworks and existing implementations (in PostGIS, Hermes, and the R package trajectories). We describe how MovingPandas avoids limitations of Simple Feature-based movement data models commonly used to handle trajectories in the GIS world. Finally, we present the current state of the MovingPandas implementation and demonstrate its use in stand-alone Python scripts, as well as within the context of the desktop GIS application QGIS. This work represents the first step towards a general-purpose Python library that enables researchers and practitioners in the GIS field and beyond to handle and analyse movement data more efficiently},
-language = {de},
-publisher = {Verlag der Österreichischen Akademie der Wissenschaften},
[…]
