+++ paper.md
@@ -35,7 +35,7 @@ Typical process for spatiotemporal data analysis involves multiple procedures, i
 
 - Data acquisition : In this aspect, existing software mainly provides data acquisition of basic geospatial data. `osmnx` [@boeing2017osmnx] is a Python package to download geospatial data from OpenStreetMap. Some packages provide tools to generate synthetic mobility data using standard mathematical models. `scikit-mobility` [@scikitmobility] is a library that allow for managing and generate mobility data of various formats.
 - Data preprocessing and data analysis: Spatiotemporal data in transportation field involves multiple source of data. Most existing tools are developed for specific type of data, such as trajectory data, air traffic data, etc. `MovingPandas` [@graser2019movingpandas] provide trajectory data structures and functions for data exploration and visualization on the top of `geopandas`. `PTRAIL` is a library for mobility data preprocessing especially in feature generation and trajectory interpolation. `traffic` [@Olive2019] is a toolbox for processing and analyzing air traffic data. `trackintel` [@axhausen2007definition] is a framework for spatio-temporal analysis of tracking data focusing on human mobility. `PySAL` [@rey_pysal_2010] is a family of packages that allows for advanced geospatial data science, which supports the development of high-level applications.
-- Data visualization: Apart from data processing,`MovingPandas` also support trajectory visualizations. `moveVis` [@2020moveVis] provides tools to visualize movement data and temporal changes of environmental data by creating video animations. `TraViA` [@Siebinga2021] provides tools to visualize and annotate movement data in an interactive approach.
+- Data visualization: Apart from data processing, `MovingPandas` also support trajectory visualizations. `moveVis` [@2020moveVis] provides tools to visualize movement data and temporal changes of environmental data by creating video animations. `TraViA` [@Siebinga2021] provides tools to visualize and annotate movement data in an interactive approach.
 
 The above-mentioned libraries provide preprocessing and geometric analysis functions from sepcific aspects. However, much remains to be done in dealing with the task of transforming raw spatio-temporal data into valuable insights. These tasks can hardly be fully covered by these libraries. Thus, a library with good compatibility with the existing tools and providing analysis framework for transportation spatio-temporal big data can effectively facilitate the research progress in this field.
 
