### 2022-01-28 ni1o1/transbigdata#11 @ni1o1
This example includes busstop name in Chinese, we have translate it into English so that it can display normally. [Example 2-Identifying arrival and departure information from Bus GPS data.ipynb](https://github.com/ni1o1/transbigdata/blob/main/example/Example%202-Identifying%20arrival%20and%20departure%20information%20from%20Bus%20GPS%20data.ipynb)

We have translated all docstrings, error messages and comments into English, include the Ipynb notebook examples. [Example 5-community detection for bikesharing data.ipynb](https://github.com/ni1o1/transbigdata/blob/main/example/Example%205-community%20detection%20for%20bikesharing%20data.ipynb) is also in English now.
