### 2022-01-28 #4021 @ni1o1
Hi @martinfleis @anitagraser @jGaboardi. Thank you for all the comments. We have finished our revision of the paper and the repo, details include:
1. Language Translation: We have translated all docstrings, error messages and comments into English, include the Ipynb notebook examples.
2. Adding Code Tests: Tests were added and the test coverage is now at 94%.
3. Paper Enhancing: We added two parts into the paper:
- The `state of the art` part introducing similar software in spatial-temporal data analysis field.
- The statement of target audience in the `state of need part`.
4. Other Improvements: We have added the contribution guidelines, code of conduct into the repo.
