+++ paper.bib
@@ -13,7 +13,6 @@
       title={PTRAIL -- A python package for parallel trajectory data preprocessing},
       author={Salman Haidri and Yaksh J. Haranwala and Vania Bogorny and Chiara Renso and Vinicius Prado da Fonseca and Amilcar Soares},
       year={2021},
-      doi = {2108.13202},
       eprint={2108.13202},
       url={https://arxiv.org/abs/2108.13202},
       archivePrefix={arXiv},
@@ -76,6 +75,7 @@
     year = {2021},
     note = {Google-Books-ID: Fb88EAAAQBAJ},
     keywords = {Transportation / General},
+    doi={10.1016/c2020-0-02866-5}
 }
 
 @article{harrison:2020,
