### 2022-01-28 ni1o1/transbigdata#5 @ni1o1
We have added the contribution guidelines, code of conduct into the repo.
