### 2022-01-28 ni1o1/transbigdata#8 @ni1o1
We add the statement of target audience in the `state of need` part of the manuscript, the `README.md` and the docs site.
