### 2022-01-28 ni1o1/transbigdata#7 @ni1o1
Tests are added and the test coverage is now at [94%](https://app.codecov.io/gh/ni1o1/transbigdata).
