### 2022-01-29 ni1o1/transbigdata#6 @jGaboardi
@ni1o1, as Anita points out above, Python 3.6 is deprecated by GeoPandas (and much of the scientific Python community). I would suggest testing in *at least* Python 3.7. However, is there a reason to not test in Python 3.9 or even 3.10?
