+++ paper.bib
@@ -238,3 +238,13 @@ year = {2020}
   doi          = {10.5281/zenodo.5528110},
   url          = {https://doi.org/10.5281/zenodo.5528110}
 }
+
+@article{haidri2021ptrail,
+      title={PTRAIL -- A python package for parallel trajectory data preprocessing}, 
+      author={Salman Haidri and Yaksh J. Haranwala and Vania Bogorny and Chiara Renso and Vinicius Prado da Fonseca and Amilcar Soares},
+      year={2021},
+      eprint={2108.13202},
+      url={https://arxiv.org/abs/2108.13202},
+      archivePrefix={arXiv},
+      primaryClass={cs.DC}
+}
