### 2022-01-29 618f7c8
update movingpandas citation
