+++ paper.bib
@@ -184,10 +184,10 @@ year = {2020}
 }
 
 @article{graser2019movingpandas,
-  title={MovingPandas: efficient structures for movement data in Python},
+  title={{MovingPandas: Efficient Structures for Movement Data in Python}},
   author={Graser, Anita},
   doi={10.1553/giscience2019_01_s54},
-  journal={GIForum},
+  journal={{GI_Forum - Journal for Geographic Information Science}},
   volume={1},
   pages={54--68},
   year={2019}
