### 2022-01-29 6e960ee
update/add PySAL citation
