+++ paper.bib
@@ -206,24 +206,27 @@ year = {2020}
   journal = {Journal of Open Source Software}
 }
 
-@incollection{rey_pysal_2010,
-	address = {Berlin, Heidelberg},
-	title = {{PySAL}: {A} {Python} {Library} of {Spatial} {Analytical} {Methods}},
-	isbn = {978-3-642-03647-7},
-	shorttitle = {{PySAL}},
-	url = {https://doi.org/10.1007/978-3-642-03647-7_11},
-	abstract = {This chapter describes PySAL, an open source library for spatial analysis written in the object oriented language Python. PySAL grew out of the software development activities that were part of the Center for Spatially Integrated Social Sciences Tools Project (Goodchild et al. 2000). This National Science Foundation infrastructure project had as its goals to facilitate dissemination of spatial analysis software to social sciences, to develop a library of spatial data analysis modules, to develop prototypes implementing state of the art methods, and to initiate and nurture a community of open source developers.},
-	language = {en},
-	urldate = {2022-01-09},
-	booktitle = {Handbook of {Applied} {Spatial} {Analysis}: {Software} {Tools}, {Methods} and {Applications}},
-	publisher = {Springer},
-	author = {Rey, Sergio J. and Anselin, Luc},
-	editor = {Fischer, Manfred M. and Getis, Arthur},
-	year = {2010},
-	doi = {10.1007/978-3-642-03647-7_11},
-	keywords = {Exploratory Spatial Data Analysis, Spatial Analytical Method, Spatial Data Analysis, Spatial Weight, Spatial Weight Matrice},
-	pages = {175--193},
+@article{pysal2007,
+  author={Sergio Rey and Luc Anselin},
+  title={{PySAL: A Python Library of Spatial Analytical Methods}},
+  journal={The Review of Regional Studies},
+  year=2007,
+  volume={37},
+  number={1},
+  pages={5-27},
+  keywords={Open Source; Software; Spatial},
+  url={https://rrs.scholasticahq.com/article/8285.pdf}
 }
+
+@article{Rey2021,
+  author = {Rey, S.J. and Anselin, L. and Amaral, P. and Arribas-Bel, D. and Cortes, R.X. and Gaboardi, J.D. and Kang, W. and Knaap, E. and Li, Z. and Lumnitz, S. and Oshan, T.M. and Shao, H. and Wolf, L.J.},
+  title = {{The PySAL Ecosystem: Philosophy and Implementation}},
+  journal = {Geographical Analysis},
+  year = {2021},
+  doi = {10.1111/gean.12276},
+  abstract = {PySAL is a library for geocomputation and spatial data science. Written in Python, the library has a long history of supporting novel scholarship and broadening methodological impacts far afield of academic work. Recently, many new techniques, methods of analyses, and development modes have been implemented, making the library much larger and more encompassing than that previously discussed in the literature. As such, we provide an introduction to the library as it stands now, as well as the scientific and conceptual underpinnings of its core set of components. Finally, we provide a prospective look at the library's future evolution.}
+}
+
 @software{scikitmobility,
   author       = {Pappalardo, Luca and
                   Simini, Filippo and
