+++ .github/workflows/tests.yml
@@ -16,10 +16,10 @@ jobs:
 
     steps:
     - uses: actions/checkout@v2
-    - name: Set up Python 3.6
+    - name: Set up Python 3.10
       uses: actions/setup-python@v2
       with:
-        python-version: "3.6"
+        python-version: "3.10"
 
     - name: Install dependencies
       run: |
