### 2022-01-29 ni1o1/transbigdata#14 @jGaboardi
[JOSS] moving pandas citation
I seem to have made in mistake (618f7c8e5d5aecdc3e4dce08d0b3e6a5d3ddc366) in the MovingPandas citation that omits the journal title (see line 100 [here](https://github.com/openjournals/joss-papers/blob/joss.04021/joss.04021/10.21105.joss.04021.pdf)).
