### 2022-01-29 ni1o1/transbigdata#6 @ni1o1
> @ni1o1, as Anita points out above, Python 3.6 is deprecated by GeoPandas (and much of the scientific Python community). I would suggest testing in _at least_ Python 3.7. However, is there a reason to not test in Python 3.9 or even 3.10?

I tried to setup the testing in Python3.10 but it failed to install the `fiona` dependency. The testing is now setup in Python3.9 and passed.
