### 2022-01-29 ni1o1/transbigdata#6 @jGaboardi
OK, I'm can put in PR that will test on matrix of Python 3.{7,8,9}, if you like. Let me know.

Another consideration moving forward (though not necessary for the JOSS paper review) is to provide an install via `conda-forge`. This will clear up a lot of install/dependency issues and make `transbigdata` more accessible for users.
