### 2022-01-29 ni1o1/transbigdata#14 @jGaboardi
Hmmmm. This is still an issue after 3b7e997c66cd474c797db7a0f0f4afc00a2328a8 (omission of the journal title). Maybe I'm making an obvious syntax problem?

@ni1o1 @anitagraser @martinfleis
