+++ .github/workflows/tests.yml
@@ -5,9 +5,19 @@ name: Tests
 
 on:
   push:
-    branches: [ main ]
+    branches:
+      - '*'
   pull_request:
-    branches: [ main ]
+    branches:
+      - '*'
+  schedule:
+      - cron: '59 23 * * *'
+  workflow_dispatch:
+    inputs:
+      version:
+        description: Manual Test Trigger
+        default: test
+        required: false
 
 jobs:
   build:
