+++ .github/workflows/tests.yml
@@ -21,15 +21,20 @@ on:
 
 jobs:
   build:
-
-    runs-on: ubuntu-latest
-
+    name: ${{ matrix.os }}, ${{ matrix.python-version }}
+    runs-on: ${{ matrix.os }}
+    timeout-minutes: 30
+    strategy:
+      matrix:
+        os: [ubuntu-latest]
+        python-version: ["3.7", "3.8", "3.9"]
+    
     steps:
     - uses: actions/checkout@v2
-    - name: Set up Python 3.6
+    - name: Set up Python ${{ matrix.python-version }}
       uses: actions/setup-python@v2
       with:
-        python-version: "3.6"
+        python-version: ${{ matrix.python-version }}
 
     - name: Install dependencies
       run: |
@@ -53,7 +58,7 @@ jobs:
         pip install pytest-cov
         pytest -v -r s --color=yes --cov=src --cov-append --cov-report term-missing --cov-report xml 
     
-    - name: "Upload coverage to Codecov"
+    - name: Upload coverage to Codecov
       uses: codecov/codecov-action@v2
       with:
         fail_ci_if_error: true
