+++ setup.py
@@ -36,6 +36,4 @@ setuptools.setup(
     package_dir={"": "src"},
     packages=setuptools.find_packages(where="src"),
     python_requires=">=3.7",
-	test_suite='nose.collector',
-    tests_require=['nose'],
 )
