### 2022-01-29 #4021 @jGaboardi
@martinfleis

I have completed my second pass. Once ni1o1/transbigdata#17 & ni1o1/transbigdata#18 are merged and ni1o1/transbigdata#14 is addressed, I believe all my comments will be resolved. That still leaves ni1o1/transbigdata#11 and ni1o1/transbigdata#12 that @anitagraser raised and anything else she may find.

@ni1o1, nice job responding to and addressing my concerns. Great work. Also, nice logo for package.
