### 2022-01-30 ni1o1/transbigdata#14 @jGaboardi
#19 solved the underscore problem, but now it seems that the hyphen in the journal name is rendering.

https://github.com/openjournals/joss-papers/blob/joss.04021/joss.04021/10.21105.joss.04021.pdf
