+++ src/transbigdata/__init__.py
@@ -1,5 +1,6 @@
 
-from plot_map import plot_map,plotscale,set_mapboxtoken,set_imgsavepath
+from .plot_map import plot_map,plotscale,set_mapboxtoken,set_imgsavepath
+from .CoordinatesConverter import *
 from .grids import *
 from .grids_geohash import geohash_encode,geohash_decode,geohash_togrid
 from .gisprocess import *
@@ -12,5 +13,4 @@ from .quality import *
 from .busgps import *
 from .getbusdata import *
 from .visualizion import *
-from CoordinatesConverter import *
-__version__ = '0.3.3'
+__version__ = '0.3.4'
