### 2022-01-31 a7707ca
Update test_bikedata.py
