+++ requirements.txt
@@ -3,8 +3,6 @@ geopandas
 matplotlib
 seaborn
 scipy
-plot_map >= 0.3.4
-CoordinatesConverter>=0.1.4
 networkx
 keplergl
 igraph
