+++ CONTRIBUTING.md
@@ -4,28 +4,16 @@ Whether you are a novice or experienced software developer, all contributions an
 
 ## Getting Started
 
-If you are looking to contribute to the *TransBigData* codebase, the best place to start is the [GitHub &#34;issues&#34; tab](https://github.com/ni1o1/transbigdata/issues).
-This is also a great place for filing bug reports and making suggestions for
-ways in which we can improve the code and documentation.
+If you are looking to contribute to the *TransBigData* codebase, the best place to start is the [GitHub &#34;issues&#34; tab](https://github.com/ni1o1/transbigdata/issues). This is also a great place for filing bug reports and making suggestions for ways in which we can improve the code and documentation.
 
-## Contributing to the Codebase
+## Step-by-step Instructions of Contribute
 
 The code is hosted on [GitHub](https://github.com/ni1o1/transbigdata),
 so you will need to use [Git](http://git-scm.com/) to clone the project and make
-changes to the codebase. Once you have obtained a copy of the code, you should
-create a development environment that is separate from your existing Python
-environment so that you can make and test changes without compromising your
-own work environment.
+changes to the codebase. 
 
-### Run the tests
-
-Before submitting your changes for review, make sure to check that your changes
-do not break any tests by running: ``pytest``. The tests are located in the [tests](https://github.com/ni1o1/transbigdata/tree/main/src/transbigdata/tests) folder.
-
-### Raising Pull Requests
-
-Once your changes are ready to be submitted, make sure to push your changes to
-your fork of the GitHub repo before creating a pull request.  We will review
-your changes, and might ask you to make additional changes before it is finally
-ready to merge. However, once it's ready, we will merge it, and you will have
-successfully contributed to the codebase!
+1. Fork the [TransBigData repository](https://github.com/ni1o1/transbigdata).
+2. Create a new branch from the `TransBigData` master branch.
+3. Within your forked copy, the source code of `TransBigData` is located at the `src` folder, you can make and test changes in the source code.
+4. Before submitting your changes for review, make sure to check that your changes do not break any tests by running: ``pytest``. The tests are located in the [tests](https://github.com/ni1o1/transbigdata/tree/main/src/transbigdata/tests) folder.
+4. When you are ready to submit your contribution, raise the Pull Request(PR). Once you finished your PR, all you have to do is wait for feedback from our review. We will review your changes, and might ask you to make additional changes before it is finally ready to merge. However, once it's ready, we will merge it, and you will have successfully contributed to the codebase!
