+++ CONTRIBUTING.md
@@ -16,4 +16,4 @@ changes to the codebase.
 2. Create a new branch from the `TransBigData` master branch.
 3. Within your forked copy, the source code of `TransBigData` is located at the [src](https://github.com/ni1o1/transbigdata/tree/main/src) folder, you can make and test changes in the source code.
 4. Before submitting your changes for review, make sure to check that your changes do not break any tests by running: ``pytest``. The tests are located in the [tests](https://github.com/ni1o1/transbigdata/tree/main/src/transbigdata/tests) folder.
-4. When you are ready to submit your contribution, raise the Pull Request(PR). Once you finished your PR, all you have to do is wait for feedback from our review. We will review your changes, and might ask you to make additional changes before it is finally ready to merge. However, once it's ready, we will merge it, and you will have successfully contributed to the codebase!
+4. When you are ready to submit your contribution, raise the Pull Request(PR). Once you finished your PR, the github [testing workflow](https://github.com/ni1o1/transbigdata/actions/workflows/tests.yml) will test your code. We will review your changes, and might ask you to make additional changes before it is finally ready to merge. However, once it's ready, we will merge it, and you will have successfully contributed to the codebase!
