+++ src/transbigdata/__init__.py
@@ -13,4 +13,4 @@ from .quality import *
 from .busgps import *
 from .getbusdata import *
 from .visualizion import *
-__version__ = '0.3.6ndansnisnfias'
+__version__ = '0.3.5'
