### 2022-02-20 ni1o1/transbigdata#24 @anitagraser
[JOSS] Install instructions: recommended Python versions
This is a follow-up of https://github.com/ni1o1/transbigdata/issues/6.

Please include the recommended Python version(s) in the README installation section.

xref: [openjournals/joss-reviews#4021](https://github.com/openjournals/joss-reviews/issues/4021)
