### 2022-02-20 ni1o1/transbigdata#25 @anitagraser
[JOSS] Missing dependencies after conda install
I followed the conda install instructions

> You can also install TransBigData by conda-forge, this will automaticaly solve the dependency, it can be installed with:
>
> `conda install -c conda-forge transbigdata`

and it seems that not all dependencies are resolved because I get

`ModuleNotFoundError: No module named 'scipy'`

![image](https://user-images.githubusercontent.com/590385/154849155-f1820093-9cb1-4c56-9882-201427426021.png)

xref: [openjournals/joss-reviews#4021](https://github.com/openjournals/joss-reviews/issues/4021)
