### 2022-02-20 ni1o1/transbigdata#25 @anitagraser
In my attempt to reevaluate https://github.com/ni1o1/transbigdata/issues/12, I have, so far, run

```
conda create -n transbigdata -c conda-forge python=3.9
conda activate transbigdata
conda install -c conda-forge transbigdata
conda install -c conda-forge jupyterlab
conda install -c conda-forge scipy
conda install -c conda-forge osmnx
```
