### 2022-02-20 ni1o1/transbigdata#26 @anitagraser
[JOSS] NameError: name 'o_node_id' is not defined in Example 4 notebook
o_node_id is used before it is defined:

![image](https://user-images.githubusercontent.com/590385/154851675-9360c3e6-3a3c-4897-b8f5-da9a5106c25f.png)

xref: [openjournals/joss-reviews#4021](https://github.com/openjournals/joss-reviews/issues/4021)
