### 2022-02-20 ni1o1/transbigdata#27 @anitagraser
[JOSS] ModuleNotFoundError: No module named 'plot_map' in Example 1 notebook
![image](https://user-images.githubusercontent.com/590385/154851849-1c37758b-ea74-4e60-bd3c-8246de9a8c7a.png)

xref: [openjournals/joss-reviews#4021](https://github.com/openjournals/joss-reviews/issues/4021)
