### 2022-02-20 ni1o1/transbigdata#25 @anitagraser
To run the other notebooks, users also need: keplergl and seaborn
