### 2022-02-20 ni1o1/transbigdata#25 @anitagraser
Overall, it looks like conda users should be instructed to install everything from https://github.com/ni1o1/transbigdata/blob/main/requirements.txt + osmnx
