### 2022-02-20 #4021 @anitagraser
@martinfleis: I have completed my second pass. In this context, I tested the conda install and re-run the notebooks which revealed a few more issues that should be addressed.
