### 2022-02-21 ni1o1/transbigdata#27 @ni1o1
`transbigdata` no longer depend on `plot_map`. I deleted this line of code and the problem is fixed.
