### 2022-02-21 ni1o1/transbigdata#24 @ni1o1
Thanks, I add the recommended Python versions in the README and the docs
