### 2022-02-21 ni1o1/transbigdata#25 @ni1o1
I added the `scipy` dependency into the conda recipe. The `osmnx`,`keplergl` and `seaborn` are the optional dependency that only used in the examples, I add the instructions of how to install these packages into the example ipynb.
