### 2022-02-21 903d426
Update paper.bib
fix two literature issues
