+++ paper.bib
@@ -187,7 +187,7 @@ year = {2020}
   title={{MovingPandas}: {Efficient} {Structures} for {Movement} {Data} in {Python}},
   author={Graser, Anita},
   doi={10.1553/giscience2019_01_s54},
-  journal={GI\_Forum ‒ Journal of Geographic Information Science},
+  journal={GI\_Forum ‒- Journal of Geographic Information Science},
   volume={1},
   pages={54--68},
   year={2019}
@@ -240,14 +240,4 @@ year = {2020}
   version      = {1.1.0},
   doi          = {10.5281/zenodo.5528110},
   url          = {https://doi.org/10.5281/zenodo.5528110}
-}
-
-@article{haidri2021ptrail,
-      title={PTRAIL -- A python package for parallel trajectory data preprocessing}, 
-      author={Salman Haidri and Yaksh J. Haranwala and Vania Bogorny and Chiara Renso and Vinicius Prado da Fonseca and Amilcar Soares},
-      year={2021},
-      eprint={2108.13202},
-      url={https://arxiv.org/abs/2108.13202},
-      archivePrefix={arXiv},
-      primaryClass={cs.DC}
 }
