### 2022-02-21 ni1o1/transbigdata#14 @yuanjian24
> #19 solved the underscore problem, but now it seems that the hyphen in the journal name is rendering.
>
> https://github.com/openjournals/joss-papers/blob/joss.04021/joss.04021/10.21105.joss.04021.pdf

Hi, the hyphen in the citation is now fixed:
![image](https://user-images.githubusercontent.com/50770426/154952774-11356b29-6054-48b7-9d2b-9bd4a27852d0.png)
@jGaboardi
