### 2022-02-22 d5342ac
add function regenerate_params and grid_from_params
