+++ src/transbigdata/grids.py
@@ -350,3 +350,89 @@ def grid_params_gini(data,col = ['lon','lat'],accuracy = 500,gini = 'max',gap = 
         raise Exception('Error setting of gini') 
     params = [r['lon'],r['lat'],deltalon,deltalat]
     return params
+
+def regenerate_params(grid):
+    '''
+    Regenerate gridding params from grid.
+
+    Parameters
+    -------
+    grid : GeoDataFrame
+        grids generated by transbigdata 
+                                               
+
+    Returns
+    -------
+    params : List
+        Gridding parameters (lonStart,latStart,deltaLon,deltaLat), lonStart and latStart are the lower-left coordinates, deltaLon, deltaLat are the length and width of a single grid
+
+    Examples
+    -------
+    >>> import transbigdata as tbd
+    >>> bounds = [113.6, 22.4, 113.605, 22.405]
+    >>> grid,params = tbd.rect_grids(bounds,500)
+    >>> tbd.regenerate_params(grid)
+    (113.6, 22.4, 0.004863669213934598, 0.004496605206422906)
+    '''
+    loncol_unique = grid.drop_duplicates(subset = ['LONCOL']).copy()
+    latcol_unique = grid.drop_duplicates(subset = ['LATCOL']).copy()
+    if (len(latcol_unique)==1)|(len(loncol_unique)==1):
+        raise Exception('Generate params failed, you must ensure that there are multiple LONCOL and LATCOL in the grids')
+    loncol_unique['LONCOL_gap'] = loncol_unique['LONCOL'].shift(-1)-loncol_unique['LONCOL']
+    loncol_unique['x'] = loncol_unique['geometry'].apply(lambda r:r.exterior.coords[0][0])
+    loncol_unique['x_gap'] = loncol_unique['x'].shift(-1)-loncol_unique['x']
+    deltaLon = (loncol_unique['x_gap']/loncol_unique['LONCOL_gap']).iloc[:-1].mean()
+
+    latcol_unique['LATCOL_gap'] = latcol_unique['LATCOL'].shift(-1)-latcol_unique['LATCOL']
+    latcol_unique['x'] = latcol_unique['geometry'].apply(lambda r:r.exterior.coords[0][1])
+    latcol_unique['x_gap'] = latcol_unique['x'].shift(-1)-latcol_unique['x']
+    deltaLat = (latcol_unique['x_gap']/latcol_unique['LATCOL_gap']).iloc[:-1].mean()
+
+    a = grid.iloc[:1]
+    lonstart = (a.centroid.x-a['LONCOL'].iloc[0]*deltaLon).iloc[0]
+    latstart = (a.centroid.y-a['LATCOL'].iloc[0]*deltaLat).iloc[0]
+    params = (lonstart,latstart,deltaLon,deltaLat)
+    return params
+
+
+def grid_from_params(params,location):
+    '''
+    Generate grids from params and bounds or shape.
+
+    Parameters
+    -------
+    params : List
+        Gridding parameters (lonStart,latStart,deltaLon,deltaLat), lonStart and latStart are the lower-left coordinates, deltaLon, deltaLat are the length and width of a single grid
+    location : bounds(List) or shape(GeoDataFrame)
+        Where to generate grids.
+                                               
+    Returns
+    -------
+    grid : GeoDataFrame
+        grids generated by transbigdata 
+    
+    '''
+    if type(location)==list:
+        shape = ''
+        bounds = location
+    elif type(location)==gpd.geodataframe.GeoDataFrame:
+        shape = location
+        bounds = shape.unary_union.bounds
+    else:
[…]
