+++ .github/workflows/tests.yml
@@ -45,18 +45,18 @@ jobs:
     - name: Lint with flake8
       run: |
         # stop the build if there are Python syntax errors or undefined names
-        flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
+        flake8 . --count --select E9,F63,F7,F82 --show-source --statistics
         # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
-        flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
+        flake8 . --count --exit-zero --max-complexity 10 --max-line-length 127 --statistics
     
     - name: Test docstrings
       run: |
-        pytest -v --color=yes --doctest-modules ./src/transbigdata
+        pytest -v --color yes --doctest-modules ./src/transbigdata
     
     - name: Test 
       run: |
         pip install pytest-cov
-        pytest -v -r s --color=yes --cov=src --cov-append --cov-report term-missing --cov-report xml 
+        pytest -v -r s --color yes --cov src --cov-append --cov-report term-missing --cov-report xml
     
     - name: Upload coverage to Codecov
       uses: codecov/codecov-action@v2
