### 2022-02-27 d030fa8
set CI with micromamba
