+++ .github/workflows/tests.yml
@@ -25,40 +25,40 @@ jobs:
     runs-on: ${{ matrix.os }}
     timeout-minutes: 30
     strategy:
-      matrix:
-        os: [ubuntu-latest]
-        python-version: ["3.7", "3.8", "3.9"]
+       matrix:
+         os: [ubuntu-latest]
+         environment-file:
+           - ci/37.yaml
+           - ci/38.yaml
+           - ci/39.yaml
     
     steps:
-    - uses: actions/checkout@v2
-    - name: Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v2
-      with:
-        python-version: ${{ matrix.python-version }}
-
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        pip install flake8 pytest
-        if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
-    
-    - name: Lint with flake8
-      run: |
-        # stop the build if there are Python syntax errors or undefined names
-        flake8 . --count --select E9,F63,F7,F82 --show-source --statistics
-        # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
-        flake8 . --count --exit-zero --max-complexity 10 --max-line-length 127 --statistics
-    
-    - name: Test docstrings
-      run: |
-        pytest -v --color yes --doctest-modules ./src/transbigdata
-    
-    - name: Test 
-      run: |
-        pip install pytest-cov
-        pytest -v -r s --color yes --cov src --cov-append --cov-report term-missing --cov-report xml
-    
-    - name: Upload coverage to Codecov
-      uses: codecov/codecov-action@v2
-      with:
-        fail_ci_if_error: true
+      - name: checkout repo
+         uses: actions/checkout@v2
+       
+       - name: setup micromamba
+         uses: mamba-org/provision-with-micromamba@main
+         with:
+           environment-file: ${{ matrix.environment-file }}
+           micromamba-version: 'latest'
+      
+      - name: Lint with flake8
+        run: |
+          # stop the build if there are Python syntax errors or undefined names
+          flake8 . --count --select E9,F63,F7,F82 --show-source --statistics
+          # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
+          flake8 . --count --exit-zero --max-complexity 10 --max-line-length 127 --statistics
+      
+      - name: Test docstrings
+        run: |
+          pytest -v --color yes --doctest-modules ./src/transbigdata
+      
+      - name: Test
+        run: |
+          pip install pytest-cov
+          pytest -v -r s --color yes --cov src --cov-append --cov-report term-missing --cov-report xml
+      
+      - name: Upload coverage to Codecov
+        uses: codecov/codecov-action@v2
+        with:
+          fail_ci_if_error: true
