+++ .github/workflows/tests.yml
@@ -34,13 +34,13 @@ jobs:
     
     steps:
       - name: checkout repo
-         uses: actions/checkout@v2
+        uses: actions/checkout@v2
        
-       - name: setup micromamba
-         uses: mamba-org/provision-with-micromamba@main
-         with:
-           environment-file: ${{ matrix.environment-file }}
-           micromamba-version: 'latest'
+      - name: setup micromamba
+        uses: mamba-org/provision-with-micromamba@main
+        with:
+          environment-file: ${{ matrix.environment-file }}
+          micromamba-version: 'latest'
       
       - name: Lint with flake8
         run: |
