### 2022-02-27 e2293b0
call flake8 with python -m
