+++ .github/workflows/tests.yml
@@ -44,10 +44,11 @@ jobs:
       
       - name: Lint with flake8
         run: |
+          pip install flake8
           # stop the build if there are Python syntax errors or undefined names
-          python -m flake8 . --count --select E9,F63,F7,F82 --show-source --statistics
+          flake8 . --count --select E9,F63,F7,F82 --show-source --statistics
           # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
-          python -m flake8 . --count --exit-zero --max-complexity 10 --max-line-length 127 --statistics
+          flake8 . --count --exit-zero --max-complexity 10 --max-line-length 127 --statistics
       
       - name: Test docstrings
         run: |
