### 2022-03-01 #4021 @martinfleis
Thank you @anitagraser and @jGaboardi!

@ni1o1 The submission is now almost ready to be published.

The next steps you need to do now:

- [ ] Fix the DOI of the PySAL paper (see the post above)
- [ ] Make a new tagged release (or let me know if the latest one contains all the changes) of the software, and list the version tag of the archived version here.
- [ ]  Check that the archival deposit (e.g., in Zenodo) has the correct metadata. In particular, the title and author list should be identical to those of the paper.
