+++ paper.bib
@@ -209,6 +209,7 @@ year = {2020}
 @article{pysal2007,
   author={Sergio Rey and Luc Anselin},
   title={{PySAL: A Python Library of Spatial Analytical Methods}},
+  doi = {10.1007/978-3-642-03647-7_11},
   journal={The Review of Regional Studies},
   year=2007,
   volume={37},
