+++ paper.bib
@@ -172,7 +172,6 @@ year = {2020}
 @article{Siebinga2021,
   doi = {10.21105/joss.03607},
   url = {https://doi.org/10.21105/joss.03607},
-  doi={10.21105/joss.03607},
   year = {2021},
   publisher = {The Open Journal},
   volume = {6},
