### 2022-03-02 #4021 @ni1o1
> * [x]  Fix the DOI of the PySAL paper (see the post above)
> * [x]  Make a new tagged release (or let me know if the latest one contains all the changes) of the software, and list the version tag of the archived version here.
> * [x]   Archive the reviewed software in Zenodo or a similar service
> * [x]   Check that the archival deposit (e.g., in Zenodo) has the correct metadata. In particular, the title and author list should be identical to those of the paper.
> * [x]   Please list the DOI of the archived version here.

Thank you @martinfleis. The DOI of PySAL is added. The latest release version `v0.3.7` is available and the Zenodo DOI is `10.5281/zenodo.6236507`
