### 2022-03-02 #4021 @arfon
Looks like this is failing because of the following warning:

> Warning:  Missing character: There is no ‒ (U+2012) (U+2012) in font [lmsans10-oblique]:+tlig;!
> Failing because there were warnings.

@tarleb - could you take a look at what might be going on here?
