### 2022-03-02 #4021 @xuanxu
The value here for `doi` is invalid: https://github.com/ni1o1/transbigdata/blob/main/paper.bib#L165
`doi.org/10.1111/2041-210X.13374` should be `10.1111/2041-210X.13374`
