### 2022-03-02 #4021 @martinfleis
@xuanxu I guess that `check references` should catch this.

@ni1o1 Thanks!
