### 2022-03-03 b6104ec
Update getbusdata.py
