### 2022-03-03 832da81
Update getbusdata.rst
