### 2025-06-09 #4021 @xyluo25
> [@xyluo25](https://github.com/xyluo25) hi, securing reviewers is the role of an editor, not an author. I suggest you leave the editor of your submission handle it.

Hi @martinfleis, thank you for letting me konw, I had initially assumed I needed to invite reviewers from the recommended papers. https://github.com/openjournals/joss-reviews/issues/8069#issuecomment-2812920904
